requirements.txt
pyyaml>=6.0
anthropic>=0.25.0
openai>=1.0.0
